Initial Commit
This is the initial and quite possibly only commit for TreasureGalaxyHelper. Basic functionality is here to enter in numbers by hand and see the complete set of numbers for all letters. Most of the code being uploaded is an unfinished feature to use OCR to do this automatically. This is half working.